obsidian-agent/requirements.txt

openai>=1.3.0
python-frontmatter>=1.0.0
python-dotenv>=1.0.0
click>=8.0.0
pytest>=7.0.0
black>=23.0.0
